clean up component
