+++ src/pages/token/rewards/UnbondingPanels.tsx
@@ -14,7 +14,7 @@ import { layoutMixins } from '@/styles/layoutMixins';
 import { AssetIcon } from '@/components/AssetIcon';
 import { Output, OutputType } from '@/components/Output';
 import { Panel } from '@/components/Panel';
-import { ValidatorFaviconIcon } from '@/components/ValidatorDropdown';
+import { ValidatorFaviconIcon } from '@/components/ValidatorName';
 
 import { getUnbondingDelegations } from '@/state/accountSelectors';
 import { useAppSelector } from '@/state/appTypes';
